+++ PAF/PAF.psm1
@@ -846,7 +846,7 @@ $oldProtocol = [Net.ServicePointManager]::SecurityProtocol
 $ModuleName = "PAF"
 
 # Get the installed version of the module
-$ModuleVersion = [version]"0.2.1"
+$ModuleVersion = [version]"0.2.2"
 
 # Find the latest version of the module in the PSGallery repository
 $LatestModule = Find-Module -Name $ModuleName -Repository PSGallery