+++ PAF/docs/ReleaseNotes.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 - snippet template
 - deleted an unnecessary `snippets\user` folder
+- [#7](https://github.com/voytas75/PowershellFramework/issues/7)
 
 ## [0.2.1] - 2023.08.02