+++ PAF/docs/ReleaseNotes.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.2.3] - 2023.08.03
+
+### Added
+
+- n/a
+
+### Changed
+
+- n/a
+
 ## [0.2.2] - 2023.08.03
 
 ### Added